legacy_cobol_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-legacy_cobol_env"
version = "0.1.0"
description = "Legacy COBOL migration workbench environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m legacy_cobol_env.server.app
server = "legacy_cobol_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["legacy_cobol_env", "legacy_cobol_env.server", "legacy_cobol_env.eval", "legacy_cobol_env.training"]
package-dir = { "legacy_cobol_env" = ".", "legacy_cobol_env.server" = "server", "legacy_cobol_env.eval" = "eval", "legacy_cobol_env.training" = "training" }
